organize nav components
