+++ components/header.tsx
@@ -1,5 +1,6 @@
-import {MobileNav} from './mobile-nav'
-import {MainNav} from './main-nav'
+import {MobileNav} from './nav/mobile'
+import {MainNav} from './nav/main'
+import {NavProps} from './nav/props'
 
 export function Header({
   leftLinks,
@@ -7,13 +8,7 @@ export function Header({
   publicId,
   logoAlt,
   basePath
-}: {
-  leftLinks: {name: string; href: string}[]
-  rightLinks: {name: string; href: string}[]
-  publicId: string
-  logoAlt: string
-  basePath: string
-}) {
+}: NavProps) {
   return (
     <header className="flex w-full shrink-0 items-center bg-brand-blue-default">
       <div className="py-4 px-6 items-center container flex">
