+++ components/nav/props.d.ts
@@ -1,0 +1,7 @@
+export type NavProps = {
+  leftLinks: {name: string; href: string}[]
+  rightLinks: {name: string; href: string}[]
+  publicId: string
+  logoAlt: string
+  basePath: string
+}
